requirements.txt
asgiref==3.3.1
autopep8==1.5.7
boto3==1.18.61
botocore==1.21.61
Django==3.0.14
django-cors-headers==3.7.0
django-storages==1.12.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
jmespath==0.10.0
pycodestyle==2.7.0
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
